Move self destruct to a separate module
