+++ backend/appBackend/src/jsMain/kotlin/com/gchristov/thecodinglove/JavascriptMain.kt
@@ -20,10 +20,11 @@ import com.gchristov.thecodinglove.search.PreloadSearchPubSubHandler
 import com.gchristov.thecodinglove.search.SearchHttpHandler
 import com.gchristov.thecodinglove.search.SearchModule
 import com.gchristov.thecodinglove.searchdata.SearchDataModule
+import com.gchristov.thecodinglove.selfdestruct.SelfDestructHttpHandler
+import com.gchristov.thecodinglove.selfdestruct.SelfDestructModule
 import com.gchristov.thecodinglove.slack.SlackAuthHttpHandler
 import com.gchristov.thecodinglove.slack.SlackEventHttpHandler
 import com.gchristov.thecodinglove.slack.SlackModule
-import com.gchristov.thecodinglove.slack.SlackSelfDestructHttpHandler
 import com.gchristov.thecodinglove.slack.interactivity.SlackInteractivityHttpHandler
 import com.gchristov.thecodinglove.slack.interactivity.SlackInteractivityPubSubHandler
 import com.gchristov.thecodinglove.slack.slashcommand.SlackSlashCommandHttpHandler
@@ -57,6 +58,7 @@ private fun setupDi(): Either<Throwable, Unit> {
             MonitoringDataModule.module,
             SearchModule.module,
             SearchDataModule.module,
+            SelfDestructModule.module,
             SlackModule.module,
             SlackDataModule.module,
         )
@@ -96,7 +98,7 @@ private suspend fun setupAppService(): Either<Throwable, HttpService> {
         DiGraph.inject<SlackInteractivityPubSubHandler>(),
         DiGraph.inject<SlackAuthHttpHandler>(),
         DiGraph.inject<SlackEventHttpHandler>(),
-        DiGraph.inject<SlackSelfDestructHttpHandler>(),
+        DiGraph.inject<SelfDestructHttpHandler>(),
         // Link this last so that API handlers are correctly registered
         StaticFileHttpHandler("$staticWebsiteRoot/index.html"),
     )
