+++ backend/self-destruct/build.gradle.kts
@@ -1,0 +1,13 @@
+plugins {
+    id("backend-service-plugin")
+}
+
+kotlin {
+    sourceSets {
+        val commonMain by getting {
+            dependencies {
+                api(projects.slackData)
+            }
+        }
+    }
+}
